refactor(Home): moved button example section into separate component
