+++ apps/web/src/app/globals.css
@@ -47,30 +47,3 @@ a {
   min-height: 500px;
   margin: 20px;
 }
-
-.example-button {
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  align-items: start;
-  background-color: #292f36;
-  border-radius: 6px;
-}
-
-.example-button-controller {
-  display: flex;
-  gap: 20px;
-  align-items: center;
-}
-
-.example-button-item {
-  display: flex;
-  gap: 20px;
-  align-items: center;
-}
-
-.example-button-name {
-  display: inline-block;
-  min-width: 100px;
-}
