+++ apps/web/src/components/ButtonExample/ButtonExample.module.css
@@ -1,0 +1,32 @@
+.button-example {
+  margin-top: 10px;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  align-items: start;
+  background-color: #292f36;
+  border-radius: 6px;
+}
+
+.button-example-controller {
+  margin-bottom: 20px;
+  display: flex;
+  gap: 20px;
+  align-items: center;
+}
+
+.button-example-item {
+  display: flex;
+  gap: 20px;
+  align-items: center;
+}
+
+.button-example-name {
+  display: inline-block;
+  min-width: 100px;
+}
+
+.button-selected {
+  background-color: #2374ab;
+}
