+++ apps/web/src/components/ButtonExample/index.ts
@@ -1,0 +1,1 @@
+export * from './ButtonExample'
